refactor(ffi): split windows.rs into windows/ submodules
Move windows.rs into windows/{to_core,list,focus,free_one,op}.rs.
- to_core.rs: ad_window_to_core helper (pub(crate) re-exported so
  tree/get.rs keeps its existing crate::windows::ad_window_to_core path)
- list.rs: ad_list_windows + ad_free_windows
- focus.rs: ad_focus_window
- free_one.rs: ad_free_window (single-struct free from ad_launch_app;
  Unit 5b renames to ad_release_window_fields)
- op.rs: ad_window_op

Remove now-unused convert::{window_info_to_c, free_window_info_fields}
re-exports.

No behavior or ABI change. Header symbol set identical.
