+++ crates/ffi/src/windows/free_one.rs
@@ -1,0 +1,12 @@
+use crate::convert::window::free_window_info_fields;
+use crate::types::AdWindowInfo;
+
+/// # Safety
+/// `win` must be null or point to a valid `AdWindowInfo`.
+#[no_mangle]
+pub unsafe extern "C" fn ad_free_window(win: *mut AdWindowInfo) {
+    if win.is_null() {
+        return;
+    }
+    free_window_info_fields(&mut *win);
+}
