+++ crates/ffi/src/windows/mod.rs
@@ -1,0 +1,7 @@
+pub(crate) mod focus;
+pub(crate) mod free_one;
+pub(crate) mod list;
+pub(crate) mod op;
+pub(crate) mod to_core;
+
+pub(crate) use to_core::ad_window_to_core;
